Added image service layer

+++ src/main/java/com/crewmaker/controller/ParticipationContoller.java
@@ -4,16 +4,14 @@ import com.crewmaker.entity.Event;
 import com.crewmaker.entity.Participation;
 import com.crewmaker.entity.User;
 import com.crewmaker.exception.ResourceNotFoundException;
-import com.crewmaker.model.UserProfile.UserProfileUser;
 import com.crewmaker.repository.EventRepository;
 import com.crewmaker.repository.ParticipationRepository;
 import com.crewmaker.repository.UserRepository;
-import com.crewmaker.reqbody.ApiResponse;
+import com.crewmaker.dto.response.ApiResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -35,7 +33,7 @@ public class ParticipationContoller {
     ParticipationRepository participationRepository;
 
     @GetMapping("/api/joinevent")
-    public ResponseEntity<?> joinEvent(@RequestParam int eventID){
+    public ResponseEntity<?> joinEvent(@RequestParam Long eventID){
         Event event = eventRepository.findByEventId(eventID);
         long actualParticipants = eventRepository.countAllByEventParticipationsIdEvent(event);
 
@@ -56,7 +54,7 @@ public class ParticipationContoller {
     }
 
     @GetMapping("/api/leaveevent")
-    public ResponseEntity<?> leaveEvent(@RequestParam int eventID){
+    public ResponseEntity<?> leaveEvent(@RequestParam Long eventID){
         Event event = eventRepository.findByEventId(eventID);
 
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
@@ -70,7 +68,7 @@ public class ParticipationContoller {
     }
 
     @GetMapping("/api/existsparticipation")
-    public boolean existParticipation(@RequestParam int eventID){
+    public boolean existParticipation(@RequestParam Long eventID){
         Event event = eventRepository.findByEventId(eventID);
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
         User user = userRepository.findByUsername(username)
@@ -80,7 +78,7 @@ public class ParticipationContoller {
     }
 
     @GetMapping("/api/eventParticipants")
-    public List<User> getEventParticipants(@RequestParam int eventID){
+    public List<User> getEventParticipants(@RequestParam Long eventID){
         Event event = eventRepository.findByEventId(eventID);
         return userRepository.findDistinctByUserParticipationsId_IdEvent(event);
     }
