+++ src/main/java/com/crewmaker/repository/CyclePeriodRepository.java
@@ -4,5 +4,5 @@ import com.crewmaker.entity.CyclePeriod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CyclePeriodRepository extends JpaRepository<CyclePeriod,Long> {
-    CyclePeriod findByCyclePeriodId(int id);
+    CyclePeriod findByCyclePeriodId(Long id);
 }
